Module Ppx_hash_lib.Std

This module is for use by ppx_hash, and is thus not in the interface of Base.

This module is for use by ppx_hash, and is thus not in the interface of Base.

module Hash = Base.Hash